pyproject.toml
[tool.poetry]
name = "wms-backend"
version = "0.1.0"
description = ""
authors = ["crossingstorage <it@crossingstorage.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"
django-simple-history = "^3.4.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
